Make a compound index rather than multikey index
It seems like multikey indicies basically create one index "axis" with
documents occupying multiple spots in that index. This makes it very
easy to find all documents that have any coordinate with a range,
but when querying individual axes, MongoDb reverts to a COLLSCAN
and takes a long time.

Solution:
This can be fixed by making a compound index, using each axis
as a key for a new index.
Wierd note: MongoDb seems to revert to a COLLSCAN if no bound is
given for the first key in the compound index.